feat: Functioning New Dataset and Ranking UIs
Added logic to signal the Backend, and for the Backend to store data in Session state.
Backend can define a “New Dataset” with a set of image names.
Modified FrontEnd to store loaded bitmaps in App (top level) state, so they can be retrieved instead of reloaded from disk.

Generalized code for writing bitmaps to canvases.
Modified RankingUI to display preloaded bitmaps.
Modified RankingUI to maintain an internal queue of choices, request new choices from the BackEnd, and load the appropriate choices into the visual UI, in time with the animation.
Backend can serve randomized choices to FrontEnd in the form of image names.
FrontEnd can report choices to the BackEnd, which stores them.